LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/view-timeline-get-set-range-expected.txt


FAIL Getting and setting the animation range assert_approx_equals: expected 100 +/- 0.000001 but got 0
